+++ src/Ext/Building/Body.h
@@ -55,6 +55,7 @@ public:
 		bool HasSuperWeapon(int index, bool withUpgrades) const;
 		bool HandleInfiltrate(HouseClass* pInfiltratorHouse , int moneybefore);
 		void UpdatePrimaryFactoryAI();
+		void KickOutStuckUnits();
 		virtual ~ExtData() = default;
 
 		// virtual void LoadFromINIFile(CCINIClass* pINI) override;